Add abstract class movableItems as parent for balls and skittles
