+++ plugins/robots/common/twoDModel/include/twoDModel/engine/model/constants.h
@@ -22,7 +22,7 @@ namespace twoDModel {
 const qreal lowPrecision = 0.00001;
 
 const qreal robotMass = 800;
-const QSize skittleSize(20, 20);
+const QSize skittleSize(50, 50);
 const QSize ballSize(50, 50);
 const qreal robotWidth = 50;
 const qreal robotHeight = 50;
