+++ plugins/robots/common/twoDModel/include/twoDModel/engine/model/robotModel.h
@@ -31,7 +31,7 @@ namespace twoDModel {
 
 namespace items {
 class StartPosition;
-class BallItem;
+class MovableItem;
 }
 
 namespace model {
